+++ ansible.cfg
@@ -1,6 +1,6 @@
 [defaults]
 inventory         = ./inventory/development
 vault_password_file = ./vault_pass.txt
-library = ./library/generated
+library = ./library
 module_utils = ./module_utils
 httpapi_plugins = ./httpapi_plugins
