+++ module_utils/http.py
@@ -1,17 +1,8 @@
 import time
 from ansible.module_utils.facts.timeout import TimeoutError
 
-API_PREFIX = "/api/fdm/v2"
-
 DEFAULT_PAGE_SIZE = 10
 DEFAULT_TIMEOUT = 60 * 10  # 10 minutes
-
-
-def base_headers(token):
-    return {
-        'Authorization': 'Bearer %s' % token,
-        'Content-Type': 'application/json'
-    }
 
 
 def iterate_over_pageable_resource(resource, params):
